Minor code style improvements
- Change font-size units from pt to px
- Move `components/Favicon/Auto/hooks/useFaviconGrabbers` to `services/FaviconGrabberService`
- Move `hooks/useJsonExport` to `services/JsonFileService`
- Move `hooks/useFontAwesomeIconList` to `services/FontAwesomeService`
- Remove `as` property from `FormLabel` component
- Remove unused `aspectRatio` property from `constants/tileSizes`
