+++ .eslintrc.json
@@ -2,7 +2,7 @@
 	"root": true,
 	// "parser": "@babel/eslint-parser",
 	"parserOptions": {
-		"ecmaVersion": 2021,
+		"ecmaVersion": 2022,
 		"sourceType": "module",
 		"ecmaFeatures": {
 			"jsx": true
