+++ src/components/Bookmark/Group/Header/header.module.scss
@@ -7,7 +7,7 @@
 
 	h1 {
 		font-weight: 400;
-		font-size: 12pt;
+		font-size: 16px;
 	}
 
 	&_edit {
@@ -18,7 +18,7 @@
 			width: 100%;
 			box-sizing: border-box;
 
-			font-size: 12pt;
+			font-size: 16px;
 			padding: 4px;
 		}
 	}
