--- src/components/Favicon/Auto/hooks/useCorsProxy.js
@@ -1,5 +1,0 @@
-export const fixCorsError = (url) => {
-	return `https://corsproxy.io/?${url}`;
-};
-
-export default () => ({ fixCorsError });
