+++ src/components/Favicon/Auto/index.jsx
@@ -1,13 +1,11 @@
 import React, { useEffect, useState } from 'react';
 import PropTypes from 'prop-types';
-import useFaviconGrabbers from './hooks/useFaviconGrabbers';
+import FaviconGrabberService from '../../../services/FaviconGrabberService';
 import UrlService from '../../../services/UrlService';
 
 const FaviconAuto = ({ websiteUrl, ...props }) => {
 	const [faviconUrl, setFaviconUrl] = useState(null);
 	const [isLoading, setIsLoading] = useState(true);
-
-	const { getFaviconWithChromeExtensionApi, getFaviconWithDuckDuckGo, getFaviconFromWebsite } = useFaviconGrabbers();
 
 	useEffect(() => {
 		if (!UrlService.isValid(websiteUrl)) {
@@ -15,20 +13,7 @@ const FaviconAuto = ({ websiteUrl, ...props }) => {
 			return;
 		}
 
-		Promise.any([
-			getFaviconWithChromeExtensionApi(websiteUrl),
-			getFaviconWithDuckDuckGo(websiteUrl),
-
-			getFaviconFromWebsite(websiteUrl, 'favicon.ico'),
-			getFaviconFromWebsite(websiteUrl, 'favicon.svg'),
-			getFaviconFromWebsite(websiteUrl, 'apple-touch-icon.png'),
-			getFaviconFromWebsite(websiteUrl, 'favicon-32x32.png'),
-			getFaviconFromWebsite(websiteUrl, 'favicon-48x48.png'),
-			getFaviconFromWebsite(websiteUrl, 'favicon-64x64.png'),
-			getFaviconFromWebsite(websiteUrl, 'favicon-167x167.png'),
-			getFaviconFromWebsite(websiteUrl, 'favicon-180x180.png'),
-			getFaviconFromWebsite(websiteUrl, 'favicon-192x192.png')
-		])
+		FaviconGrabberService.tryAllPossibleCases(websiteUrl)
 			.then((favicon) => {
 				setFaviconUrl(favicon);
 				// TODO: save a favicon's url in LocalStorage for quicker loading and reducing an amount of requests
