+++ src/components/Form/IconPicker/icon-picker.module.scss
@@ -33,7 +33,7 @@
 		align-items: center;
 		text-align: center;
 		gap: 8px;
-		font-size: 10pt;
+		font-size: 14px;
 		cursor: pointer;
 		padding: 10px 5px;
 
@@ -56,6 +56,6 @@
 	}
 
 	&__icon {
-		font-size: 20pt;
+		font-size: 26px;
 	}
 }
