+++ src/components/Form/IconPicker/index.jsx
@@ -2,13 +2,14 @@ import React, { useMemo, useState } from 'react';
 import classNames from 'classnames';
 import PropTypes from 'prop-types';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-import useFontAwesomeIconList from '../../../hooks/useFontAwesomeIconList';
+import FontAwesomeService from '../../../services/FontAwesomeService';
 import styles from './icon-picker.module.scss';
+
+const icons = FontAwesomeService.getIconList();
 
 const FormIconPicker = ({ selectedIconName, onPick }) => {
 	const [searchValue, setSearchValue] = useState('');
 	const [selectedIcon, setSelectedIcon] = useState(selectedIconName);
-	const icons = useFontAwesomeIconList();
 
 	const handleSearchInputChange = (event) => {
 		setSearchValue(event.target.value);
