+++ src/components/Form/Label/index.jsx
@@ -3,17 +3,16 @@ import classNames from 'classnames';
 import PropTypes from 'prop-types';
 import styles from './label.module.scss';
 
-const FormLabel = ({ label, required, error, vertical, as = 'a', children }) => {
-	const WrapperComponent = as;
+const FormLabel = ({ label, required, error, vertical, children }) => {
 	return (
 		<div className={classNames(styles.formLabel, { [styles.formLabelError]: error })}>
-			<WrapperComponent className={classNames(styles.formLabelBody, { [styles.formLabelVertical]: vertical })}>
+			<label className={classNames(styles.formLabelBody, { [styles.formLabelVertical]: vertical })}>
 				<span className={styles.formLabelTitle}>
 					{label}
 					{required && <span className={styles.formLabelRequiredMark}>*</span>}
 				</span>
 				{children}
-			</WrapperComponent>
+			</label>
 			{error && (
 				<span className={styles.formLabelErrorMessage} role="alert">
 					{error.message}
@@ -28,7 +27,6 @@ FormLabel.propTypes = {
 	required: PropTypes.bool,
 	error: PropTypes.shape({ message: PropTypes.string.isRequired }),
 	vertical: PropTypes.bool,
-	as: PropTypes.string,
 	children: PropTypes.element
 };
