+++ src/components/Form/RadioGroup/index.jsx
@@ -4,7 +4,7 @@ import FormLabel from '../Label';
 import styles from './radio-group.module.scss';
 
 const RadioGroup = ({ children, ...labelProps }) => (
-	<FormLabel as="div" vertical {...labelProps}>
+	<FormLabel vertical {...labelProps}>
 		<div className={styles.radioGroup}>{children}</div>
 	</FormLabel>
 );
