+++ src/components/Modal/modal.module.scss
@@ -29,7 +29,7 @@ $back-bg: #202020;
 		padding: $content-padding * 1.2 $content-padding;
 		background-color: $back-bg;
 		font-weight: 500;
-		font-size: 13pt;
+		font-size: 18px;
 	}
 
 	&__close-button {
