+++ src/components/Sidebar/sidebar.module.scss
@@ -60,7 +60,7 @@
 		&-icon {
 			width: $sidebar-covered-width;
 			min-width: $sidebar-covered-width;
-			font-size: 15pt;
+			font-size: 20px;
 		}
 
 		&-label {
